Introduce internal config, remove system class path from private configuration
