+++ exonum-java-binding-core/rust/integration_tests/tests/service_runtime.rs
@@ -4,7 +4,7 @@ extern crate java_bindings;
 
 use exonum_testkit::TestKitBuilder;
 use integration_tests::vm::get_fakes_classpath;
-use java_bindings::{Config, JavaServiceRuntime, PrivateConfig, PublicConfig};
+use java_bindings::{Config, InternalConfig, JavaServiceRuntime, PrivateConfig, PublicConfig};
 
 const TEST_SERVICE_MODULE_NAME: &str =
     "com.exonum.binding.fakes.services.service.TestServiceModule";
@@ -17,16 +17,21 @@ fn bootstrap() {
 
     let private_config = PrivateConfig {
         user_parameters: Vec::new(),
-        system_class_path: get_fakes_classpath(),
         service_class_path: "".to_string(),
         log_config_path: "".to_owned(),
         port: 6000,
     };
 
-    let service_runtime = JavaServiceRuntime::get_or_create(Config {
-        public_config,
-        private_config,
-    });
+    let service_runtime = JavaServiceRuntime::get_or_create(
+        Config {
+            public_config,
+            private_config,
+        },
+        InternalConfig {
+            system_class_path: get_fakes_classpath(),
+            system_lib_path: "".to_string(),
+        },
+    );
 
     let mut testkit = TestKitBuilder::validator()
         .with_service(service_runtime.service_proxy())
