+++ exonum-java-binding-core/rust/src/runtime/java_service_runtime.rs
@@ -3,11 +3,12 @@ use exonum::helpers::fabric::{CommandExtension, Context, ServiceFactory};
 use jni::{self, JavaVM};
 
 use std::env;
+use std::fs;
 use std::sync::{Arc, Once, ONCE_INIT};
 
 use proxy::{JniExecutor, ServiceProxy};
 use runtime::cmd::{Finalize, GenerateTemplate, Run};
-use runtime::config::{self, Config, PrivateConfig};
+use runtime::config::{self, Config, InternalConfig, PrivateConfig};
 use utils::{current_directory, join_paths, unwrap_jni};
 use MainExecutor;
 
@@ -30,11 +31,11 @@ impl JavaServiceRuntime {
     /// Creates new runtime from provided config or returns the one created earlier.
     ///
     /// There can be only one `JavaServiceRuntime` instance at a time.
-    pub fn get_or_create(config: Config) -> Self {
+    pub fn get_or_create(config: Config, internal: InternalConfig) -> Self {
         unsafe {
             // Initialize runtime if it wasn't created before.
             JAVA_SERVICE_RUNTIME_INIT.call_once(|| {
-                let java_vm = Self::create_java_vm(&config.private_config);
+                let java_vm = Self::create_java_vm(&config.private_config, internal);
                 let executor = MainExecutor::new(Arc::new(java_vm));
                 let service_proxy = Self::create_service(
                     &config.public_config.module_name,
@@ -64,7 +65,7 @@ impl JavaServiceRuntime {
     /// # Panics
     ///
     /// - If user specified invalid additional JVM parameters.
-    fn create_java_vm(config: &PrivateConfig) -> JavaVM {
+    fn create_java_vm(config: &PrivateConfig, internal_config: InternalConfig) -> JavaVM {
         let mut args_builder = jni::InitArgsBuilder::new().version(jni::JNIVersion::V8);
 
         for param in &config.user_parameters {
@@ -72,11 +73,18 @@ impl JavaServiceRuntime {
             args_builder = args_builder.option(&option);
         }
 
-        let class_path = join_paths(&[&config.system_class_path, &config.service_class_path]);
+        let class_path = join_paths(&[
+            &internal_config.system_class_path,
+            &config.service_class_path,
+        ]);
 
         args_builder = args_builder.option(&format!("-Djava.class.path={}", class_path));
-        args_builder =
-            args_builder.option(&format!("-Djava.library.path={}", absolute_library_path()));
+        if internal_config.system_lib_path != "" {
+            args_builder = args_builder.option(&format!(
+                "-Djava.library.path={}",
+                internal_config.system_lib_path
+            ));
+        }
         args_builder = args_builder.option(&format!(
             "-Dlog4j.configurationFile={}",
             config.log_config_path
@@ -113,6 +121,26 @@ fn absolute_library_path() -> String {
         current_directory
     };
     library_path.to_string_lossy().into_owned()
+}
+
+fn system_classpath() -> String {
+    let mut jars = Vec::new();
+    let jars_directory = {
+        let mut current_directory = current_directory();
+        current_directory.push("lib/java");
+        current_directory
+    };
+    for entry in fs::read_dir(jars_directory).expect("Could not read java classes directory") {
+        let file = entry.unwrap();
+        if file.file_type().unwrap().is_file() {
+            jars.push(file.path());
+        } else {
+            continue;
+        }
+    }
+
+    let jars = jars.iter().map(|p| p.to_str().unwrap());
+    env::join_paths(jars).unwrap().into_string().unwrap()
 }
 
 /// Panics if `_JAVA_OPTIONS` environmental variable is set.
@@ -157,7 +185,11 @@ impl ServiceFactory for JavaServiceFactory {
                 .clone()
                 .try_into()
                 .expect("Invalid EJB configuration format.");
-            JavaServiceRuntime::get_or_create(config)
+            let internal_config = InternalConfig {
+                system_class_path: system_classpath(),
+                system_lib_path: absolute_library_path(),
+            };
+            JavaServiceRuntime::get_or_create(config, internal_config)
         };
 
         Box::new(runtime.service_proxy().clone())
